super saver fixes - still unintended save needs to be fixed

+++ HKWeb/src/com/hk/web/action/core/catalog/image/UploadSuperSaverImageAction.java
@@ -23,10 +23,12 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 import org.apache.log4j.Logger;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang.StringUtils;
 
+@Component
 public class UploadSuperSaverImageAction extends BaseAction {
     FileBean fileBean;
     List<SuperSaverImage> superSaverImages;
@@ -77,43 +79,49 @@ public class UploadSuperSaverImageAction extends BaseAction {
         } catch (IOException ioe) {
             logger.error("Error while uploading super saver image: " + ioe);
         } finally {
+	        if(imageFile != null){
             FileUtils.deleteQuietly(imageFile);
+	        }
         }
         addRedirectAlertMessage(new SimpleMessage("Error while uploading image!"));
         return new ForwardResolution("/pages/uploadSuperSaverImage.jsp");
     }
 
-    public Resolution editSuperSaverImageSettings() {
-        if (superSaverImages != null) {
-            for (SuperSaverImage superSaverImage : superSaverImages) {
-                Product superSaverProduct = superSaverImage.getProduct();
+	public Resolution editSuperSaverImageSettings() {
+		if (superSaverImages != null) {
+			for (SuperSaverImage superSaverImage : superSaverImages) {
+				Product superSaverProduct = superSaverImage.getProduct();
+				if (superSaverProduct != null) {
+					//check whether combo exists or not
+					Combo combo = comboDao.getComboById(superSaverProduct.getId());
+					if (combo != null) {
+						//while place this check once main image id logic is implemented
+						/*if (superSaverImage.isHidden() == Boolean.TRUE) {
+												if (superSaverImage.isMainImage() == Boolean.TRUE) {
+													logger.debug("main image cannot be set as hidden!");
+													addRedirectAlertMessage(new SimpleMessage("MAIN IMAGE CANNOT BE SET AS HIDDEN!"));
+													return new ForwardResolution("/pages/manageSuperSaverImages.jsp");
+												}
+											}*/
+						superSaverImage.setMainImage(Boolean.TRUE);
 
-                //check whether combo exists or not
-                Combo combo = comboDao.getComboById(superSaverProduct.getId());
-                if (combo != null) {
-                    //while place this check once main image id logic is implemented
-                    /*if (superSaverImage.isHidden() == Boolean.TRUE) {
-                        if (superSaverImage.isMainImage() == Boolean.TRUE) {
-                            logger.debug("main image cannot be set as hidden!");
-                            addRedirectAlertMessage(new SimpleMessage("MAIN IMAGE CANNOT BE SET AS HIDDEN!"));
-                            return new ForwardResolution("/pages/manageSuperSaverImages.jsp");
-                        }
-                    }*/
-                    superSaverImage.setMainImage(Boolean.TRUE);
-
-                    String altText = superSaverImage.getAltText();
-                    String productName = superSaverProduct.getName();
-                    superSaverImage.setUrl(productName);
-                    superSaverImage.setAltText(StringUtils.isNotBlank(altText) ? altText : productName);
-                    superSaverImageService.saveSuperSaverImage(superSaverImage);
-                } else {
-                    addRedirectAlertMessage(new SimpleMessage("No combo exists with the specified id! Kindly enter a valid combo id."));
-                    return new ForwardResolution("/pages/uploadSuperSaverImage.jsp");
-                }
-            }
-        }
-        return new RedirectResolution(SuperSaversAction.class);
-    }
+						String altText = superSaverImage.getAltText();
+						String productName = superSaverProduct.getName();
+						superSaverImage.setUrl(productName);
+						superSaverImage.setAltText(StringUtils.isNotBlank(altText) ? altText : productName);
+						superSaverImageService.saveSuperSaverImage(superSaverImage);
+					} else {
+						addRedirectAlertMessage(new SimpleMessage("No combo exists with the specified id! Kindly enter a valid combo id."));
+						return new RedirectResolution(UploadSuperSaverImageAction.class, "manageSuperSaverImages");
+					}
+				} else {
+					addRedirectAlertMessage(new SimpleMessage("No combo exists with the specified id! Kindly enter a valid combo id."));
+					return new RedirectResolution(UploadSuperSaverImageAction.class, "manageSuperSaverImages");
+				}
+			}
+		}
+		return new RedirectResolution(SuperSaversAction.class);
+	}
 
     public FileBean getFileBean() {
         return fileBean;
